01_Areas/Codebases/ailcc/automations/crew_ai/requirements.txt
crewai
langchain
websocket-client
python-dotenv
